# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Participant & Session API",
    "description": "REST API for managing participants and sessions",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://cse341-week3-4-9ppj.onrender.com"
    }
  ],
  "paths": {
    "/participants": {
      "get": {
        "summary": "Get all participants",
        "responses": {
          "200": { "description": "Success" },
          "500": { "description": "Server error" }
        }
      },
      "post": {
        "summary": "Create participant",
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Validation error" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/participants/{id}": {
      "get": {
        "summary": "Get participant by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Success" },
          "404": { "description": "Not found" }
        }
      },
      "put": {
        "summary": "Update participant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Updated" },
          "400": { "description": "Validation error" },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "summary": "Delete participant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Deleted" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/sessions": {
      "get": {
        "summary": "Get all sessions",
        "responses": {
          "200": { "description": "Success" }
        }
      },
      "post": {
        "summary": "Create session",
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Validation error" }
        }
      }
    },
    "/sessions/{id}": {
      "get": {
        "summary": "Get session by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Success" },
          "404": { "description": "Not found" }
        }
      },
      "put": {
        "summary": "Update session",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Updated" }
        }
      },
      "delete": {
        "summary": "Delete session",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Deleted" }
        }
      }
    }
  }
}